requirements.txt
# Core dependencies
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1
pandas==2.0.1
pyyaml==6.0
requests==2.31.0
flask==2.3.2
flask-cors==4.0.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-dotenv==1.0.0
click==8.1.3
tqdm==4.65.0

# FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
xtb-python==0.0.1

# Development dependencies
black==23.3.0
flake8==6.0.0
mypy==1.3.0
pre-commit==3.3.2
